# perf(plugin-search): reindex collections in parallel, up to 80% faster