--- a/test/plugin-search/seed/index.ts
+++ b/test/plugin-search/seed/index.ts
@@ -14,30 +14,6 @@ export const seed = async (payload: Payload): Promise<boolean> => {
       req,
     })
 
-    // seed 1000 pages
-    for (let i = 0; i < 1000; i++) {
-      await payload.create({
-        collection: 'pages',
-        data: {
-          title: `Test page title ${i}`,
-          _status: 'published',
-        },
-        req,
-      })
-    }
-
-    // seed 1000 posts
-    for (let i = 0; i < 1000; i++) {
-      await payload.create({
-        collection: 'posts',
-        data: {
-          title: `Test post title ${i}`,
-          _status: 'published',
-        },
-        req,
-      })
-    }
-
     return true
   } catch (err) {
     console.error(err)