@lukebennett 2025-09-19
Is there a particular reason for having everything part of the same transaction, rather than one transaction per collection, or even per batch?