@jacobsfletch 2025-09-19
> Is there a particular reason for having everything part of the same transaction, rather than one transaction per collection, or even per batch?

I'm open to it. Trying to think of scenarios where all or nothing indexing might be preferred or even required.